start open and bid
